pyproject.toml
[project]
name = "soloqueue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.7",
    "html2text>=2025.4.15",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "langchain-core>=1.2.9",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "python-frontmatter>=1.1.0",
    "python-multipart>=0.0.22",
    "tiktoken>=0.12.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/soloqueue"]
